Moved DevTools test serializers into a separate folder
