+++ packages/react-devtools-shared/src/__tests__/__serializers__/inspectedElementSerializer.js
